{"schema":"decodex/commit/1","summary":"Split MCP observability projections","authority":"XY-1115"}
