+++ apps/decodex/src/mcp/observability/projection/lane.rs
@@ -1,0 +1,79 @@
+use serde_json::{self, Value};
+
+use crate::mcp::observability::projection::review;
+
+pub(in crate::mcp) fn mcp_public_lane_inspect_resource(report: Value) -> Value {
+	serde_json::json!({
+		"schema": "decodex.mcp.lane_inspect/1",
+		"projectId": report.get("projectId").cloned().unwrap_or(Value::Null),
+		"issue": report.get("issue").cloned().unwrap_or(Value::Null),
+		"runId": report.get("runId").cloned().unwrap_or(Value::Null),
+		"matchedRunCount": report.get("matchedRunCount").cloned().unwrap_or(Value::Null),
+		"runs": mcp_public_lane_inspect_runs(report.get("runs"))
+	})
+}
+
+pub(in crate::mcp) fn mcp_public_post_review_lane(lane: &Value) -> Value {
+	serde_json::json!({
+		"project_id": lane.get("project_id").cloned().unwrap_or(Value::Null),
+		"issue_id": lane.get("issue_id").cloned().unwrap_or(Value::Null),
+		"issue_identifier": lane.get("issue_identifier").cloned().unwrap_or(Value::Null),
+		"issue_state": lane.get("issue_state").cloned().unwrap_or(Value::Null),
+		"classification": lane.get("classification").cloned().unwrap_or(Value::Null),
+		"reason": lane.get("reason").cloned().unwrap_or(Value::Null),
+		"pr_url": lane.get("pr_url").cloned().unwrap_or(Value::Null),
+		"pr_state": lane.get("pr_state").cloned().unwrap_or(Value::Null),
+		"review_decision": lane.get("review_decision").cloned().unwrap_or(Value::Null),
+		"mergeable": lane.get("mergeable").cloned().unwrap_or(Value::Null),
+		"check_state": lane.get("check_state").cloned().unwrap_or(Value::Null),
+		"unresolved_review_threads": lane
+			.get("unresolved_review_threads")
+			.cloned()
+			.unwrap_or(Value::Null),
+		"shadowed_by_current_lane": lane
+			.get("shadowed_by_current_lane")
+			.cloned()
+			.unwrap_or(Value::Null),
+		"readback_warning": lane.get("readback_warning").cloned().unwrap_or(Value::Null),
+		"readback_root_cause": lane.get("readback_root_cause").cloned().unwrap_or(Value::Null),
+		"loop_review": lane
+			.get("loop_status")
+			.and_then(review::mcp_loop_review_status_from_loop_status)
+			.map(review::mcp_public_review_status)
+			.unwrap_or(Value::Null)
+	})
+}
+
+pub(super) fn mcp_public_post_review_lanes(lanes: Option<&Value>) -> Vec<Value> {
+	lanes.and_then(Value::as_array).into_iter().flatten().map(mcp_public_post_review_lane).collect()
+}
+
+fn mcp_public_lane_inspect_runs(runs: Option<&Value>) -> Vec<Value> {
+	runs.and_then(Value::as_array).into_iter().flatten().map(mcp_public_lane_inspect_run).collect()
+}
+
+fn mcp_public_lane_inspect_run(run: &Value) -> Value {
+	serde_json::json!({
+		"projectId": run.get("projectId").cloned().unwrap_or(Value::Null),
+		"issueId": run.get("issueId").cloned().unwrap_or(Value::Null),
+		"issueIdentifier": run.get("issueIdentifier").cloned().unwrap_or(Value::Null),
+		"runId": run.get("runId").cloned().unwrap_or(Value::Null),
+		"attemptNumber": run.get("attemptNumber").cloned().unwrap_or(Value::Null),
+		"status": run.get("status").cloned().unwrap_or(Value::Null),
+		"attemptStatus": run.get("attemptStatus").cloned().unwrap_or(Value::Null),
+		"phase": run.get("phase").cloned().unwrap_or(Value::Null),
+		"waitReason": run.get("waitReason").cloned().unwrap_or(Value::Null),
+		"currentOperation": run.get("currentOperation").cloned().unwrap_or(Value::Null),
+		"laneControlNextAction": run
+			.get("laneControlNextAction")
+			.cloned()
+			.unwrap_or(Value::Null),
+		"laneControlConditions": run
+			.get("laneControlConditions")
+			.cloned()
+			.unwrap_or_else(|| serde_json::json!([])),
+		"lastEventType": run.get("lastEventType").cloned().unwrap_or(Value::Null),
+		"lastEventAt": run.get("lastEventAt").cloned().unwrap_or(Value::Null),
+		"eventCount": run.get("eventCount").cloned().unwrap_or(Value::Null)
+	})
+}
