+++ apps/decodex/src/mcp/observability/projection/resources.rs
@@ -1,0 +1,127 @@
+use serde_json::{self, Value};
+
+use crate::mcp::{
+	DEFAULT_MCP_STATUS_LIMIT, McpError,
+	observability::projection::{lane, protocol, review, runs},
+};
+
+pub(in crate::mcp) fn mcp_status_live_resource(snapshot: Value) -> Value {
+	serde_json::json!({
+		"schema": "decodex.mcp.status_live/1",
+		"project_id": snapshot.get("project_id").cloned().unwrap_or(Value::Null),
+		"status_source": snapshot.get("status_source").cloned().unwrap_or(Value::Null),
+		"run_limit": snapshot.get("run_limit").cloned().unwrap_or(Value::Null),
+		"current_lanes": runs::mcp_run_activity_summaries(snapshot.get("current_lanes")),
+		"recent_runs": runs::mcp_run_activity_summaries(snapshot.get("recent_runs")),
+		"post_review_lanes": lane::mcp_public_post_review_lanes(snapshot.get("post_review_lanes"))
+	})
+}
+
+pub(in crate::mcp) fn mcp_activity_tail_resource(snapshot: Value) -> Value {
+	let limit = snapshot
+		.get("run_limit")
+		.and_then(Value::as_u64)
+		.and_then(|limit| usize::try_from(limit).ok())
+		.unwrap_or(DEFAULT_MCP_STATUS_LIMIT);
+	let mut activity = Vec::new();
+
+	for run in runs::mcp_all_runs(&snapshot).into_iter().take(limit) {
+		activity.push(runs::mcp_run_activity_summary(run));
+	}
+
+	serde_json::json!({
+		"schema": "decodex.mcp.activity_tail/1",
+		"project_id": snapshot.get("project_id").cloned().unwrap_or(Value::Null),
+		"activity": activity
+	})
+}
+
+pub(in crate::mcp) fn mcp_public_lane_control_readback_resource(readback: Value) -> Value {
+	serde_json::json!({
+		"schema": "decodex.mcp.lane_control_readback/1",
+		"project_id": readback.get("project_id").cloned().unwrap_or(Value::Null),
+		"read_only": readback.get("read_only").cloned().unwrap_or(Value::Null),
+		"mutating_tools": readback.get("mutating_tools").cloned().unwrap_or_else(|| serde_json::json!([])),
+		"current_lanes": runs::mcp_run_activity_summaries(readback.get("current_lanes")),
+		"recent_runs": runs::mcp_run_activity_summaries(readback.get("recent_runs")),
+		"post_review_lanes": lane::mcp_public_post_review_lanes(readback.get("post_review_lanes"))
+	})
+}
+
+pub(in crate::mcp) fn mcp_run_resource(
+	snapshot: &Value,
+	run_id: &str,
+	kind: &str,
+) -> Result<Value, McpError> {
+	let Some(run) = runs::mcp_find_run(snapshot, run_id) else {
+		return Err(McpError::resource_not_found());
+	};
+	let value = match kind {
+		"events" => serde_json::json!({
+			"schema": "decodex.mcp.run_events/1",
+			"project_id": snapshot.get("project_id").cloned().unwrap_or(Value::Null),
+			"run_id": run.get("run_id").cloned().unwrap_or(Value::Null),
+			"issue_id": run.get("issue_id").cloned().unwrap_or(Value::Null),
+			"issue_identifier": run.get("issue_identifier").cloned().unwrap_or(Value::Null),
+			"event_count": run.get("event_count").cloned().unwrap_or(Value::Null),
+			"last_event_type": run.get("last_event_type").cloned().unwrap_or(Value::Null),
+			"last_event_at": run.get("last_event_at").cloned().unwrap_or(Value::Null),
+			"last_protocol_activity_at": run
+				.get("last_protocol_activity_at")
+				.cloned()
+				.unwrap_or(Value::Null)
+		}),
+		"protocol_activity" => serde_json::json!({
+			"schema": "decodex.mcp.protocol_activity/1",
+			"project_id": snapshot.get("project_id").cloned().unwrap_or(Value::Null),
+			"run_id": run.get("run_id").cloned().unwrap_or(Value::Null),
+			"protocol_activity": protocol::mcp_public_protocol_activity(run),
+			"event_count": run.get("event_count").cloned().unwrap_or(Value::Null),
+			"last_event_type": run.get("last_event_type").cloned().unwrap_or(Value::Null)
+		}),
+		"child_agent_activity" => serde_json::json!({
+			"schema": "decodex.mcp.child_agent_activity/1",
+			"project_id": snapshot.get("project_id").cloned().unwrap_or(Value::Null),
+			"run_id": run.get("run_id").cloned().unwrap_or(Value::Null),
+			"child_agent_activity": run.get("child_agent_activity").cloned().unwrap_or(Value::Null)
+		}),
+		"progress_diagnostics" => serde_json::json!({
+			"schema": "decodex.mcp.progress_diagnostics/1",
+			"project_id": snapshot.get("project_id").cloned().unwrap_or(Value::Null),
+			"run_id": run.get("run_id").cloned().unwrap_or(Value::Null),
+			"phase": run.get("phase").cloned().unwrap_or(Value::Null),
+			"run_phase": run.get("run_phase").cloned().unwrap_or(Value::Null),
+			"current_operation": run.get("current_operation").cloned().unwrap_or(Value::Null),
+			"last_progress_at": run.get("last_progress_at").cloned().unwrap_or(Value::Null),
+			"progress_diagnostic": run.get("progress_diagnostic").cloned().unwrap_or(Value::Null),
+			"suspected_stall": run.get("suspected_stall").cloned().unwrap_or(Value::Null)
+		}),
+		_ => unreachable!("MCP run resource kind is selected by static match arms"),
+	};
+
+	Ok(value)
+}
+
+pub(in crate::mcp) fn mcp_pr_review_state_resource(snapshot: Value) -> Value {
+	let review_lanes = lane::mcp_public_post_review_lanes(snapshot.get("post_review_lanes"));
+	let current_lane_reviews = runs::mcp_current_lane_runs(&snapshot)
+		.into_iter()
+		.filter_map(|run| {
+			let review = review::mcp_loop_review_status(run)?;
+
+			Some(serde_json::json!({
+				"run_id": run.get("run_id").cloned().unwrap_or(Value::Null),
+				"issue_id": run.get("issue_id").cloned().unwrap_or(Value::Null),
+				"issue_identifier": run.get("issue_identifier").cloned().unwrap_or(Value::Null),
+				"review": review::mcp_public_review_status(review)
+			}))
+		})
+		.collect::<Vec<_>>();
+
+	serde_json::json!({
+		"schema": "decodex.mcp.pr_review_state/1",
+		"project_id": snapshot.get("project_id").cloned().unwrap_or(Value::Null),
+		"post_review_lanes": review_lanes,
+		"current_lane_reviews": current_lane_reviews
+	})
+}
