+++ apps/decodex/src/mcp/observability/projection/review.rs
@@ -1,0 +1,44 @@
+use serde_json::{self, Value};
+
+pub(super) fn mcp_public_phase_acceptance_status(acceptance: &Value) -> Value {
+	serde_json::json!({
+		"phase": acceptance.get("phase").cloned().unwrap_or(Value::Null),
+		"decision": acceptance.get("decision").cloned().unwrap_or(Value::Null),
+		"reason_code": acceptance.get("reason_code").cloned().unwrap_or(Value::Null),
+		"objective_covered": acceptance.get("objective_covered").cloned().unwrap_or(Value::Null),
+		"effective_delta_present": acceptance
+			.get("effective_delta_present")
+			.cloned()
+			.unwrap_or(Value::Null),
+		"non_goal_passed": acceptance.get("non_goal_passed").cloned().unwrap_or(Value::Null),
+		"validation_passed": acceptance.get("validation_passed").cloned().unwrap_or(Value::Null),
+		"next_action": acceptance.get("next_action").cloned().unwrap_or(Value::Null)
+	})
+}
+
+pub(super) fn mcp_public_review_status(review: &Value) -> Value {
+	serde_json::json!({
+		"phase": review.get("phase").cloned().unwrap_or(Value::Null),
+		"status": review.get("status").cloned().unwrap_or(Value::Null),
+		"checkpoint": review
+			.get("checkpoint")
+			.map(mcp_public_review_checkpoint_status)
+			.unwrap_or(Value::Null)
+	})
+}
+
+pub(super) fn mcp_loop_review_status(run_or_lane: &Value) -> Option<&Value> {
+	run_or_lane.get("loop_status").and_then(mcp_loop_review_status_from_loop_status)
+}
+
+pub(super) fn mcp_loop_review_status_from_loop_status(loop_status: &Value) -> Option<&Value> {
+	loop_status.get("review").filter(|review| review.is_object())
+}
+
+fn mcp_public_review_checkpoint_status(checkpoint: &Value) -> Value {
+	serde_json::json!({
+		"round": checkpoint.get("round").cloned().unwrap_or(Value::Null),
+		"nonclean_rounds": checkpoint.get("nonclean_rounds").cloned().unwrap_or(Value::Null),
+		"updated_at": checkpoint.get("updated_at").cloned().unwrap_or(Value::Null)
+	})
+}
