+++ apps/decodex/src/mcp/observability/projection/runs.rs
@@ -1,0 +1,82 @@
+use std::collections::BTreeSet;
+
+use serde_json::{self, Value};
+
+use crate::mcp::observability::projection::{autonomy, protocol, review};
+
+pub(super) fn mcp_run_activity_summaries(runs: Option<&Value>) -> Vec<Value> {
+	runs.and_then(Value::as_array).into_iter().flatten().map(mcp_run_activity_summary).collect()
+}
+
+pub(super) fn mcp_all_runs(snapshot: &Value) -> Vec<&Value> {
+	let mut runs = Vec::new();
+	let mut seen_run_ids = BTreeSet::new();
+
+	for key in ["current_lanes", "recent_runs"] {
+		if let Some(items) = snapshot.get(key).and_then(Value::as_array) {
+			for (index, run) in items.iter().enumerate() {
+				let run_key = run
+					.get("run_id")
+					.and_then(Value::as_str)
+					.map(str::to_owned)
+					.unwrap_or_else(|| format!("{key}:{index}"));
+
+				if seen_run_ids.insert(run_key) {
+					runs.push(run);
+				}
+			}
+		}
+	}
+
+	runs
+}
+
+pub(super) fn mcp_current_lane_runs(snapshot: &Value) -> Vec<&Value> {
+	snapshot.get("current_lanes").and_then(Value::as_array).into_iter().flatten().collect()
+}
+
+pub(super) fn mcp_find_run<'a>(snapshot: &'a Value, run_id: &str) -> Option<&'a Value> {
+	mcp_all_runs(snapshot)
+		.into_iter()
+		.find(|run| run.get("run_id").and_then(Value::as_str) == Some(run_id))
+}
+
+pub(in crate::mcp) fn mcp_run_activity_summary(run: &Value) -> Value {
+	serde_json::json!({
+		"run_id": run.get("run_id").cloned().unwrap_or(Value::Null),
+		"issue_id": run.get("issue_id").cloned().unwrap_or(Value::Null),
+		"issue_identifier": run.get("issue_identifier").cloned().unwrap_or(Value::Null),
+		"attempt_number": run.get("attempt_number").cloned().unwrap_or(Value::Null),
+		"status": run.get("status").cloned().unwrap_or(Value::Null),
+		"attempt_status": run.get("attempt_status").cloned().unwrap_or(Value::Null),
+		"phase": run.get("phase").cloned().unwrap_or(Value::Null),
+		"run_phase": run.get("run_phase").cloned().unwrap_or(Value::Null),
+		"wait_reason": run.get("wait_reason").cloned().unwrap_or(Value::Null),
+		"current_operation": run.get("current_operation").cloned().unwrap_or(Value::Null),
+		"lane_control_next_action": run
+			.get("lane_control_next_action")
+			.cloned()
+			.unwrap_or(Value::Null),
+		"event_count": run.get("event_count").cloned().unwrap_or(Value::Null),
+		"last_event_type": run.get("last_event_type").cloned().unwrap_or(Value::Null),
+		"last_event_at": run.get("last_event_at").cloned().unwrap_or(Value::Null),
+		"last_protocol_activity_at": run
+			.get("last_protocol_activity_at")
+			.cloned()
+			.unwrap_or(Value::Null),
+		"last_progress_at": run.get("last_progress_at").cloned().unwrap_or(Value::Null),
+		"protocol_activity": protocol::mcp_public_protocol_activity(run),
+		"child_agent_activity": run.get("child_agent_activity").cloned().unwrap_or(Value::Null),
+		"progress_diagnostic": run.get("progress_diagnostic").cloned().unwrap_or(Value::Null),
+		"phase_acceptance": run
+			.get("phase_acceptance")
+			.map(review::mcp_public_phase_acceptance_status)
+			.unwrap_or(Value::Null),
+		"autonomy": autonomy::mcp_public_autonomy_status(run),
+		"loop_review": run
+			.get("loop_status")
+			.and_then(review::mcp_loop_review_status_from_loop_status)
+			.map(review::mcp_public_review_status)
+			.unwrap_or(Value::Null)
+	})
+}
